% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prv_adjust_interaction_ci.R
\name{adjust_interaction_ci}
\alias{adjust_interaction_ci}
\title{Adjust an interaction term's confidence interval (CI)}
\usage{
adjust_interaction_ci(modelobj, data = NULL)
}
\arguments{
\item{modelobj}{The model to adjust.}

\item{data}{The data used to fit the model. If \code{NULL} (default), tries to
detect what was used in \code{modelobj}.}
}
\value{
A named List. \verb{$lwr} and \verb{$upr} contain named Numeric vectors that
are the lower and upper 95\% confidence interval, respectively. The names in
each vector are the names of the terms from the model.
}
\description{
Simply uses the adjusted coefficients and SEs to recalculate the 95\%
confidence interval.
}
