% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prv_adjust_interaction_coef.R
\name{adjust_interaction_coef}
\alias{adjust_interaction_coef}
\title{Adjust an interaction term's coefficients}
\usage{
adjust_interaction_coef(modelobj, data)
}
\arguments{
\item{modelobj}{(Object) A model object.}

\item{data}{(Dataframe) The data used to fit the model.}
}
\value{
A Numeric vector, where the names are all of the terms from the
model, and the values are the adjusted coefficients.
}
\description{
The adjustment is done by summing all of the coefficients that go into a term.
For example, if a model has the term "SexMale", then it will simply get the
coefficient it already has. However, the term "SexMale:EducationPrimary" would
be adjusted as "SexMale:EducationPrimary + SexMale + EducationPrimary".
}
\keyword{internal}
