% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_adjust_interaction_model.R
\name{adjust_interaction_model}
\alias{adjust_interaction_model}
\title{Return an adjusted \code{summary()} table for an interaction model}
\usage{
adjust_interaction_model(modelobj, data, exponentiate = FALSE)
}
\arguments{
\item{modelobj}{(Object) A model object.}

\item{data}{(Dataframe) The data used to fit the model.}

\item{exponentiate}{(Logical) If \code{TRUE}, exponentiates the coefficient and confidence interval.}
}
\value{
A data frame.
}
\description{
Adjusts the coefficients and CIs of any interaction terms in the model.
}
\examples{
# cancer_modified is a built-in dataset provided with the `coefixr` package.
my_model <- lm(status ~ age + sex * ph.ecog + sex * wt.loss, data = cancer_modified)

rownames(summary(my_model)$coefficients)
adjust_interaction_model(my_model, cancer_modified, exponentiate = TRUE)

}
