% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prv_adjust_interaction_se.R
\name{adjust_interaction_se}
\alias{adjust_interaction_se}
\title{Adjust an interaction term's Standard Error (SE)}
\usage{
adjust_interaction_se(modelobj)
}
\arguments{
\item{modelobj}{(Object) The model to adjust.}
}
\value{
A named Numeric vector, where the names are all of the terms from the
model, and the values are the adjusted SEs.
}
\description{
The adjustment is done by 1) retrieving the variance-covariance table for the
model and multiplying the off-diagonal cells (covariances) by 2, then
2) selecting the cells that are relevant to the current term, then
3) summing all those together and taking their square root.
}
\keyword{internal}
