% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prv_build_missing_terms.R
\name{build_missing_terms}
\alias{build_missing_terms}
\title{Build all of the missing terms in a model}
\usage{
build_missing_terms(modelobj, data)
}
\arguments{
\item{modelobj}{(Object) A model object.}

\item{data}{(Dataframe) The data used to fit the model.}
}
\value{
A named List.
- \code{original_terms} has the original terms given by R.
- \code{missing_terms} are the terms that were not originally given, which involve the reference levels of factors in interactions.
- \code{filled_terms} is the union of \code{original_terms} and \code{missing_terms}.
- \code{with_singletons} also includes the reference terms outside interactions.
- \code{reference_levels} is the reference level of each Factor used in the model.
}
\description{
R does not return interactions of covariates that involve the reference
levels of factors. This function finds all of the interactions in a model,
crosses all of the levels together, and returns that information and more.
}
