% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prv_helpers.R
\name{named_to_df}
\alias{named_to_df}
\title{Convert a named vector into a two-column data frame}
\usage{
named_to_df(vec, value = "value", name = "covar")
}
\arguments{
\item{vec}{(Vector) A named vector.}

\item{value}{(Character) The name of the column that will hold the values of \code{vec}.}

\item{name}{(Character) The name of the column that will hold the names of \code{vec}.}
}
\value{
A data frame.
}
\description{
Convert a named vector into a two-column data frame
}
