% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_adjust_interaction_model.R
\name{adjust_interaction_model}
\alias{adjust_interaction_model}
\title{Calculate adjusted coefficients and CIs for models with interactions}
\usage{
adjust_interaction_model(
  modelobj,
  data,
  exponentiate = FALSE,
  add.global.p = FALSE,
  digits.n = 2,
  digits.p = 3
)
}
\arguments{
\item{modelobj}{(Object) A model object.}

\item{data}{(Dataframe) The data used to fit the model.}

\item{exponentiate}{(Logical) If \code{TRUE}, exponentiates the coefficient and confidence interval.}

\item{add.global.p}{(Logical) If \code{TRUE}, adds a global p-value for each covariate.}

\item{digits.n}{(Logical) Number of digits to round coefficients and confidence intervals to.}

\item{digits.p}{(Logical) Number of digits to round p-values to. Also handles very small ("<0.001") and large (">0.999") p-values.}
}
\value{
A data frame with these columns:
\describe{
\item{\strong{covar}}{The covariate.}
\item{\strong{ref}}{\code{TRUE} marks the reference levels of covariates.}
\item{\strong{ref.intx}}{\code{TRUE} marks interactions that involve the reference level of a covariate.}
\item{\strong{global.p}}{The global p-value of the covariate. Column is omitted if \code{add.global.p = FALSE}.}
\item{\strong{p.value}}{The p-value of the covariate.}
\item{\strong{log_ci.95lwr} or \strong{exp_ci.95lwr}}{Lower 95\% confidence interval of the coefficient. Exponentiated if \code{exponentiate = TRUE}.}
\item{\strong{log_coef} or \strong{exp_coef}}{The coefficient. Exponentiated if \code{exponentiate = TRUE}.}
\item{\strong{log_ci.95upr} or \strong{exp_ci.95upr}}{Upper 95\% confidence interval of the coefficient. Exponentiated if \code{exponentiate = TRUE}.}
}
}
\description{
Adjusts the coefficients and CIs of any interaction terms in the model.
}
\examples{
# cancer_modified is a dataset provided with the `coefixr` package.
my_model <- lm(status ~ inst + age + sex * ph.ecog + sex * wt.loss, data = cancer_modified)

# Unexponentiated coefficients.
adjust_interaction_model(my_model, cancer_modified)

# To get unrounded numbers, set digits.n and digits.p to Inf.
adjust_interaction_model(my_model, cancer_modified, digits.n = Inf, digits.p = Inf)

# Note that column names change when `exponentiate = TRUE`.
adjust_interaction_model(my_model, cancer_modified, exponentiate = TRUE)

# Global p-values are provided by `car::Anova()`, just like `gtsummary::add_global_p()`.
adjust_interaction_model(my_model, cancer_modified, add.global.p = TRUE)

}
